Move search controls under screens folder
Refs #366.
